requirements.txt
python-dotenv>=1.0.1
python-telegram-bot>=21.0,<22.0
